+++ src/components/visualizations/table/fixtures/arithmericMeasures.ts
@@ -121,8 +121,6 @@ export const TABLE_HEADERS_AM: Execution.IMeasureHeaderItem[] = [
     },
     {
         measureHeaderItem: {
-            uri: '/am1uri',
-            identifier: 'am1id',
             localIdentifier: 'am1',
             name: 'AM1(M1,M2)',
             format: '#,##0'
@@ -130,8 +128,6 @@ export const TABLE_HEADERS_AM: Execution.IMeasureHeaderItem[] = [
     },
     {
         measureHeaderItem: {
-            uri: '/am2uri',
-            identifier: 'am2id',
             localIdentifier: 'am2',
             name: 'AM2(AM1,M2)',
             format: '#,##0'
@@ -139,8 +135,6 @@ export const TABLE_HEADERS_AM: Execution.IMeasureHeaderItem[] = [
     },
     {
         measureHeaderItem: {
-            uri: '/am3uri',
-            identifier: 'am3id',
             localIdentifier: 'am3',
             name: 'AM3(M0,M2)',
             format: '#,##0'
@@ -148,8 +142,6 @@ export const TABLE_HEADERS_AM: Execution.IMeasureHeaderItem[] = [
     },
     {
         measureHeaderItem: {
-            uri: '/dam1uri',
-            identifier: 'dam1id',
             localIdentifier: 'dam1',
             name: 'DAM1(AM1)',
             format: '#,##0'
@@ -196,8 +188,6 @@ export const EXECUTION_RESPONSE_AM: Execution.IExecutionResponse = {
                             },
                             {
                                 measureHeaderItem: {
-                                    uri: '/am1uri',
-                                    identifier: 'am1id',
                                     localIdentifier: 'am1',
                                     name: 'AM1(M1,M2)',
                                     format: '#,##0'
@@ -205,8 +195,6 @@ export const EXECUTION_RESPONSE_AM: Execution.IExecutionResponse = {
                             },
                             {
                                 measureHeaderItem: {
-                                    uri: '/am2uri',
-                                    identifier: 'am2id',
                                     localIdentifier: 'am2',
                                     name: 'AM2(AM1,M2)',
                                     format: '#,##0'
@@ -214,8 +202,6 @@ export const EXECUTION_RESPONSE_AM: Execution.IExecutionResponse = {
                             },
                             {
                                 measureHeaderItem: {
-                                    uri: '/am3uri',
-                                    identifier: 'am3id',
                                     localIdentifier: 'am3',
                                     name: 'AM3(M0,A2)',
                                     format: '#,##0'
@@ -223,8 +209,6 @@ export const EXECUTION_RESPONSE_AM: Execution.IExecutionResponse = {
                             },
                             {
                                 measureHeaderItem: {
-                                    uri: '/dam1uri',
-                                    identifier: 'dam1id',
                                     localIdentifier: 'dam1',
                                     name: 'DAM1(AM1)',
                                     format: '#,##0'