+++ src/components/visualizations/utils/drilldownEventing.ts
@@ -5,20 +5,27 @@ import { AFM } from '@gooddata/typings';
 import * as CustomEvent from 'custom-event';
 import * as Highcharts from 'highcharts';
 import * as invariant from 'invariant';
-import { VisElementType, VisType, VisualizationTypes } from '../../../constants/visualizationTypes';
+import {
+    ChartElementType,
+    ChartType,
+    VisType,
+    VisualizationTypes
+} from '../../../constants/visualizationTypes';
 import {
     IDrillEvent,
     DrillEventContext,
     IDrillEventContextGroup,
     IDrillEventIntersectionElement,
-    ILegacyDrillIntersection
+    IDrillEventContextPoint,
+    IDrillEventContextTable,
+    IDrillPoint
 } from '../../../interfaces/DrillEvents';
 import { OnFiredDrillEvent } from '../../../interfaces/Events';
 import { TableRowForDrilling } from '../../../interfaces/Table';
 import { isComboChart, isHeatmap, isTreemap } from './common';
 
 export interface IHighchartsPointObject extends Highcharts.PointObject {
-    drillIntersection: ILegacyDrillIntersection[];
+    drillIntersection: IDrillEventIntersectionElement[];
     z?: number; // is missing in HCH's interface
     value?: number; // is missing in HCH's interface
 }
@@ -36,7 +43,7 @@ export interface ICellDrillEvent {
     columnIndex: number;
     rowIndex: number;
     row: TableRowForDrilling;
-    intersection: ILegacyDrillIntersection[];
+    intersection: IDrillEventIntersectionElement[];
 }
 
 export interface IDrillConfig {
@@ -44,7 +51,7 @@ export interface IDrillConfig {
     onFiredDrillEvent: OnFiredDrillEvent;
 }
 
-export function getClickableElementNameByChartType(type: VisType): VisElementType {
+export function getClickableElementNameByChartType(type: VisType): ChartElementType {
     switch (type) {
         case VisualizationTypes.LINE:
         case VisualizationTypes.AREA:
@@ -59,7 +66,6 @@ export function getClickableElementNameByChartType(type: VisType): VisElementTyp
         case VisualizationTypes.DONUT:
         case VisualizationTypes.FUNNEL:
             return 'slice';
-        case VisualizationTypes.TABLE:
         case VisualizationTypes.HEATMAP:
             return 'cell';
         default:
@@ -81,58 +87,52 @@ function fireEvent(onFiredDrillEvent: OnFiredDrillEvent, data: IDrillEvent, targ
     }
 }
 
-function normalizeIntersectionElements(intersection: ILegacyDrillIntersection[]): IDrillEventIntersectionElement[] {
-    return intersection.map(({ id, title, value, name, uri, identifier }) => {
-       const intersection: IDrillEventIntersectionElement = {
-           id,
-           title: title || value as string || name
-       };
-       if (uri || identifier) {
-           intersection.header = { uri, identifier };
-       }
-       return intersection;
-    });
-}
-
 function composeDrillContextGroup(
     { points }: IHighchartsChartDrilldownEvent,
-    chartType: VisType
+    chartType: ChartType
 ): IDrillEventContextGroup {
+    const contextPoints: IDrillPoint[] = points.map((point: IHighchartsPointObject) => {
+        return {
+            x: point.x,
+            y: point.y,
+            intersection: point.drillIntersection
+        };
+    });
     return {
         type: chartType,
         element: 'label',
-        points: points.map((point: IHighchartsPointObject) => {
-            return {
-                x: point.x,
-                y: point.y,
-                intersection: normalizeIntersectionElements(point.drillIntersection)
-            };
-        })
+        points: contextPoints
     };
 }
 
 function composeDrillContextPoint(
     { point }: IHighchartsChartDrilldownEvent,
-    chartType: VisType
-): DrillEventContext {
-    const zProp = isNaN(point.z) ? {} : { z: point.z };
-    const valueProp = (isTreemap(chartType) || isHeatmap(chartType)) ? { value: point.value.toString() } : {};
-    const xyProp = isTreemap(chartType) ? {} : {
-        x: point.x,
-        y: point.y
-    };
-    return {
+    chartType: ChartType
+): IDrillEventContextPoint {
+    const context: IDrillEventContextPoint = {
         type: chartType,
         element: getClickableElementNameByChartType(chartType),
-        ...xyProp,
-        ...zProp,
-        ...valueProp,
-        intersection: normalizeIntersectionElements(point.drillIntersection)
+        intersection: point.drillIntersection
     };
+
+    if (!isTreemap(chartType)) {
+        context.x = point.x;
+        context.y = point.y;
+    }
+
+    if (!isNaN(point.z)) {
+        context.z = point.z;
+    }
+
+    if (isTreemap(chartType) || isHeatmap(chartType)) {
+        context.value =  point.value.toString();
+    }
+
+    return context;
 }
 
 const chartClickDebounced = debounce((drillConfig: IDrillConfig, event: IHighchartsChartDrilldownEvent,
-                                      target: EventTarget, chartType: VisType) => {
+                                      target: EventTarget, chartType: ChartType) => {
     const { afm, onFiredDrillEvent } = drillConfig;
 
     let usedChartType = chartType;
@@ -155,24 +155,47 @@ const chartClickDebounced = debounce((drillConfig: IDrillConfig, event: IHighcha
 export function chartClick(drillConfig: IDrillConfig,
                            event: IHighchartsChartDrilldownEvent,
                            target: EventTarget,
-                           chartType: VisType) {
+                           chartType: ChartType) {
     chartClickDebounced(drillConfig, event, target, chartType);
 }
 
 export function cellClick(drillConfig: IDrillConfig, event: ICellDrillEvent, target: EventTarget) {
     const { afm, onFiredDrillEvent } = drillConfig;
     const { columnIndex, rowIndex, row, intersection } = event;
+
+    const drillContext: IDrillEventContextTable = {
+        type: VisualizationTypes.TABLE,
+        element: 'cell',
+        columnIndex,
+        rowIndex,
+        row,
+        intersection
+    };
     const data: IDrillEvent = {
         executionContext: afm,
-        drillContext: {
-            type: VisualizationTypes.TABLE,
-            element: getClickableElementNameByChartType(VisualizationTypes.TABLE),
-            columnIndex,
-            rowIndex,
-            row,
-            intersection: normalizeIntersectionElements(intersection)
-        }
+        drillContext
     };
 
     fireEvent(onFiredDrillEvent, data, target);
+}
+
+export function createDrillIntersectionElement(
+    id: string,
+    title: string,
+    uri?: string,
+    identifier?: string
+): IDrillEventIntersectionElement {
+    const element: IDrillEventIntersectionElement = {
+        id: id || '',
+        title: title || ''
+    };
+
+    if (uri || identifier) {
+        element.header = {
+            uri: uri || '',
+            identifier: identifier || ''
+        };
+    }
+
+    return element;
 }